feat(component): create shared PortfolioCard
