+++ src/pages/index.js
@@ -1,38 +1,16 @@
 import { Link, StaticQuery, graphql } from 'gatsby'
 
-import { GatsbyImage } from 'gatsby-plugin-image'
 import Layout from '../components/layout'
 import React from 'react'
 import Seo from '../components/seo'
 import styled from 'styled-components'
+import PortfolioCard from '../components/portfolioCard'
 
 const IndexWrapper = styled.main`
   align-items: center;
   display: flex;
   flex-direction: column;
   max-width: ${({ theme: { maxWidth } }) => maxWidth};
-`
-
-const PortfolioCard = styled.article`
-  align-items: center;
-  border: 2px solid #ffffff;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  overflow: hidden;
-  transition: all
-    ${({
-      theme: {
-        transitions: { transitionSpeed },
-      },
-    }) => transitionSpeed};
-
-  &:hover {
-    background-color: rgba(255, 255, 255, 0.2);
-    h2 {
-      color: blue;
-    }
-  }
 `
 
 const PortfolioContainer = styled.section`
@@ -56,28 +34,6 @@ const PortfolioContainer = styled.section`
     theme.mq('lg')`
     grid-template-columns: 1fr 1fr 1fr;
   `}
-`
-
-const PortfolioImgContainer = styled.div`
-  width: 100%;
-`
-
-const PortfolioTitleContainer = styled.div`
-  align-items: center;
-  display: flex;
-  flex: 1 0 auto;
-  justify-content: center;
-`
-
-const PortfolioTitleText = styled.h2`
-  color: #ffffff;
-  font-size: 1rem;
-  padding: 1.5rem;
-  text-align: center;
-`
-
-const StyledLink = styled(Link)`
-  display: flex;
 `
 
 const HOMEPAGE_QUERY = graphql`
@@ -127,23 +83,13 @@ const IndexPage = () => (
             allMarkdownRemark.edges.map(({ node }) => {
               if (node.frontmatter.homepage === 'yes') {
                 return (
-                  <StyledLink to={`/portfolio-items${node.frontmatter.slug}`}>
-                    <PortfolioCard key={node.frontmatter.slug}>
-                      <PortfolioImgContainer>
-                        <GatsbyImage
-                          image={
-                            node.frontmatter.hero.childImageSharp
-                              .gatsbyImageData
-                          }
-                        />
-                      </PortfolioImgContainer>
-                      <PortfolioTitleContainer>
-                        <PortfolioTitleText>
-                          {node.frontmatter.title}
-                        </PortfolioTitleText>
-                      </PortfolioTitleContainer>
-                    </PortfolioCard>
-                  </StyledLink>
+                  <PortfolioCard
+                    image={
+                      node.frontmatter.hero.childImageSharp.gatsbyImageData
+                    }
+                    slug={node.frontmatter.slug}
+                    title={node.frontmatter.title}
+                  />
                 )
               } else {
                 return null
